feat: add ExpressionExtensions.cs
